separo pantalla de perfil en distintos componentes para manejar mejor las estadisticas
